// typescript-projects/utility-backups/src20250512/interface/types.ts
export type Flatten<T> = T extends Record<string, unknown>
  ? UnionToIntersection<
      {
        [K in keyof DeepRequired<T>]: K extends string
          ? T[K] extends Record<string, unknown>
            ? {
                [P in `${K}.${keyof Flatten<T[K]> & string}`]: Flatten<
                  T[K]
                >[P extends `${K}.${infer R}` ? R : never];
              } & { [P in K]: T[K] }
            : { [P in K]: T[K] }
          : never;
      }[keyof T]
    >
  : never;

export type UnionToIntersection<T> = (
  T extends unknown ? (x: T) => void : never
) extends (x: infer R) => void
  ? R
  : never;

/** Recursively make all properties of a map/object required */
export type DeepRequired<T> = {
  [K in keyof T]-?: Exclude<T[K], undefined> extends Record<string, unknown>
    ? DeepRequired<T[K]>
    : T[K];
};
// ...
export type TPaymentData =
  | TEbillsAirtimePaymentData
  | TEbillsInternetPaymentData
  | TEbillsBetPaymentData
  | TEbillsCableTvPaymentData
  | TEbillsElectricityPaymentData;

// ################## EBILLS PROVIDER START ##################
export type TEbillsAirtimePaymentData = {
  provider: ProviderId;
  phoneNumber: string;
  merchant: string;
  amount: number; // `amount` in tx doc root
};

export type TEbillsInternetPaymentData = {
  provider: ProviderId;
  phoneNumber: string;
  merchant: string;
  variationId: string;
};

export type TEbillsBetPaymentData = {
  provider: ProviderId;
  customerId: string;
  merchant: string;
  amount: number; // `amount` in tx doc root
};

export type TEbillsCableTvPaymentData = {
  provider: ProviderId;
  cardNumber: string;
  merchant: string;
  variationId: string;
};

export type TEbillsElectricityPaymentData = {
  provider: ProviderId;
  merchant: string;
  customerId: string;
  amount: number; // `amount` in tx doc root
};
// ...
export type TProcessData = {
  paymentData: TPaymentData;
  uid: string;
  appTxHash: string;
  nairaAdminDocPath: string;
  txPath: string;
  provider: BaseProvider;
};

/** Data expected to come into the status-processing endpoint service class */
export type TStatusData = {
  provider: BaseProvider;
  queryStatus: AppTxStatus;
  reference?: string;
  txPath?: string;
};

/** Data for getting document refs of docs to be written/updated during utility processing. */
export type TStatusDocsAccessData = {
  uid: string;
  /** hash of naira withdrawal tx, which also serves as tx doc ID */
  appTxHash: string;
  /** full path to the related naira_admin doc */
  nairaAdminDocPath: string;
};